+++ lib/commit-meat.rb
@@ -9,9 +9,22 @@ module CommitMeat
   def self.run
     message_file = ARGV[0]
     message = File.read(message_file).strip
-    puts message.green
-    Tester.new(message)
-    puts 'erm'
+
+    tester = Tester.new(message)
+
+    tester.test
+    if tester.failed?
+      fail_with_message(tester.failure_messages)
+    end
+  end
+
+  private
+
+  def self.fail_with_message(messages)
+    puts "There's NO MEAT - did not commit.".red
+    messages.each do
+      puts message.yellow
+    end
     exit 1
   end